Set library path on some packages to find CUDA libraries in /run/opengl-driver.
This is to avoid relying on LD_LIBRARY_PATH for finding the driver libraries in /run/opengl-driver(-32)/lib.

I found these packages by searching for "cuda" in nixpkgs and guessing which packages may need such patching, usually by checking the code if they dlopen any driver libraries such as libcuda or libnv<something>.

Confirmed that Blender can use CUDA without LD_LIBRARY_PATH and it couldn't before. Did not test cudnn and nccl, but checked that the library path was applied.

Note that not every package that uses cudatoolkit needs to be patched, only if it directly loads driver libraries. Some packages
already handle the problem by directly depending on nvidia_x11 one way or another (not sure if this is a good thing) where this
patching is presumably unnecessary.